Cast keysym to an int in the debug printf

+++ generic/tkBind.c
@@ -4201,7 +4201,7 @@ HandleEventGenerate(
 		return TCL_ERROR;
 	    }
 printf("key value is %s\n", value);
-printf("keysym is %d\n", keysym); fflush(stdout);
+printf("keysym is %d\n", (int)keysym); fflush(stdout);
 	    TkpSetKeycodeAndState(tkwin, keysym, &event.general);
 	    if (event.general.xkey.keycode == 0) {
 		Tcl_SetObjResult(interp, Tcl_ObjPrintf("no keycode for keysym \"%s\"", value));
